+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/UnimplementedMessageBridgeSpec.kt
@@ -33,4 +33,14 @@ class UnimplementedMessageBridgeSpec : BridgeCommon() {
         val mockFunction: (Any) -> Unit = {}
         unimplementedMessageBridge.sendJsonToHostApp("", mockFunction, mockFunction)
     }
+
+    @Test(expected = MiniAppSdkException::class)
+    fun `getHostAppThemeColors should throw MiniAppSdkException when it is not implemented`() {
+        unimplementedMessageBridge.getHostAppThemeColors(mock(), mock())
+    }
+
+    @Test(expected = MiniAppSdkException::class)
+    fun `getIsDarkMode should throw MiniAppSdkException when it is not implemented`() {
+        unimplementedMessageBridge.getIsDarkMode(mock(), mock())
+    }
 }