+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/helper/MessageBridgeHelper.kt
@@ -1,12 +1,15 @@
 package com.rakuten.tech.mobile.testapp.helper
 
 import android.app.Activity
+import android.content.res.Configuration
 import androidx.core.app.ActivityCompat
+import com.rakuten.tech.mobile.miniapp.analytics.MAAnalyticsInfo
 import com.rakuten.tech.mobile.miniapp.errors.MiniAppAccessTokenError
 import com.rakuten.tech.mobile.miniapp.errors.MiniAppPointsError
 import com.rakuten.tech.mobile.miniapp.js.MessageToContact
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import com.rakuten.tech.mobile.miniapp.js.chat.ChatBridgeDispatcher
+import com.rakuten.tech.mobile.miniapp.js.hostenvironment.HostAppThemeColors
 import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
 import com.rakuten.tech.mobile.miniapp.js.userinfo.Points
 import com.rakuten.tech.mobile.miniapp.js.userinfo.TokenData
@@ -81,6 +84,31 @@ fun getMessageBridge(
                 activity.showToastMessage(message)
             }
         }
+    }
+
+    override fun getHostAppThemeColors(
+        onSuccess: (themeColor: HostAppThemeColors) -> Unit,
+        onError: (message: String) -> Unit
+    ) {
+        onSuccess(AppSettings.instance.colorTheme)
+    }
+
+    override fun getIsDarkMode(
+        onSuccess: (isDarkMode: Boolean) -> Unit,
+        onError: (message: String) -> Unit
+    ) {
+        when (activity.resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK) {
+            Configuration.UI_MODE_NIGHT_NO -> onSuccess(false) // Night mode is not active, we're using the light theme.
+            Configuration.UI_MODE_NIGHT_YES -> onSuccess(true) // Night mode is active, we're using dark theme.
+        }
+    }
+
+    override fun didReceiveMAAnalytics(
+        analyticsInfo: MAAnalyticsInfo,
+        onSuccess: (message: String) -> Unit,
+        onError: (message: String) -> Unit
+    ) {
+        FileUtils.saveMiniAppAnalyticLogs(activity, analyticsInfo)
     }
 }