+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ## CHANGELOG
 
+### 5.3.0 (xxxx-xx-xx)
+**SDK**
+- **Feature:** Added `MiniAppMessageBridge.getHostAppThemeColors` in MiniAppMessageBridge for retrieving primary and secondary theme color.
+- **Feature:** Added `MiniAppMessageBridge.getIsDarkMode` in MiniAppMessageBridge for retrieving dark mode status.
+- **Feature:** Added `MiniApp.unzipBundle` to load MiniApp from bundle. Host app can unzip a MiniApp bundle from asset folder into MiniApp file system.
+- **Feature:** Added `MiniAppView.loadFromBundle` to load MiniApp from already extracted MiniApp bundle.
+- **Feature:** Added `MiniAppMessageBridge.didReceiveMAAnalytics` to receive analytics from MiniApp.
+
+
+**Sample App**
+- **Feature:** Added implementation of `MiniAppMessageBridge.getHostAppThemeColors`.
+- **Feature:** Added qa settings to test primary and secondary color.
+- **Feature:** Added Deeplink Support to change settings keys by QR code.
+- **Feature:** Added implementation of `MiniAppMessageBridge.getIsDarkMode`.
+- **Feature:** Added `MiniApp` button in Feature page to load a MiniApp from bundle.
+- **Feature:** Added `Logs` in `QA` page to show the analytics events come from miniapp.
+
 ### 5.2.1 (2023-05-23)
 **SDK**
 - **Fix:** Support In-app Purchase for old sdk implementation.